コード整理。m_separateByBlendShape　オプション(default: trueが前と同じ挙動)追加
