--- Assets/MeshesIntegrated.mesh.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 48cadfe261b960249b58f8eac1fe35d4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4300000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
