--- Assets/VRM/Editor/SkinnedMeshUtility/SkinnedMeshUtility.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d66a91f4b51e42a6b165fe39fdb48d64
-timeCreated: 1559541677
