+++ Assets/VRM/Editor/SkinnedMeshUtility/VrmMeshIntegratorWizard.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: 648920943c26bde4b920119bdbb38f70
-timeCreated: 1518194696
-licenseType: Free
+guid: 0c6fe206e704d2148be9fd6ede8ca4f0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
