+++ Assets/VRM/Editor/SkinnedMeshUtility/VrmPrefabUtility.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: 744f66c7013807b4cae1d38d2ecfff38
-timeCreated: 1518503829
-licenseType: Free
+guid: 6d3e24cadf1393841988d0c53830b67e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
